# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.HTTPError:
    properties:
      error:
        type: string
    type: object
...
  models.CashFlowStatement:
    properties:
      at_beginning_of_year:
        type: string
      at_end_of_year:
        type: string
      cff_dividends_paid:
        type: string
      cff_loans_received:
        type: string
      cff_net_financing_cash_flow:
        type: string
      cff_proceeds_from_stocks_bonds_issuance_or_contributed_capital:
        type: string
      cff_repayments_of_lease_obligations:
        type: string
      cff_repayments_of_loans_made:
        type: string
      cff_subsidies_grants_donations_received:
        type: string
      cfi_acquisition_of_fixed_assets_intangible_assets:
        type: string
      cfi_acquisition_of_stocks_shares:
        type: string
      cfi_dividends_received:
        type: string
      cfi_interest_received:
        type: string
      cfi_loans_made:
        type: string
      cfi_net_investing_cash_flow:
        type: string
      cfi_repayments_of_loans_received:
        type: string
      cfi_sale_proceeds_from_fixed_assets_intangible_assets:
        type: string
      cfi_sale_proceeds_from_stocks_shares:
        type: string
      cfo_dm_cash_paid_to_suppliers_employees:
        type: string
      cfo_dm_cash_received_from_customers:
        type: string
      cfo_dm_extra_items_cash_flow:
        type: string
      cfo_dm_income_taxes_paid:
        type: string
      cfo_dm_interest_paid:
        type: string
      cfo_dm_net_operating_cash_flow:
        type: string
      cfo_dm_operating_cash_flow:
        type: string
      cfo_dm_other_cash_received_paid:
        type: string
      cfo_im_extra_items_cash_flow:
        type: string
      cfo_im_income_before_changes_in_working_capital:
        type: string
      cfo_im_income_before_income_taxes:
        type: string
      cfo_im_income_taxes_paid:
        type: string
      cfo_im_interest_paid:
        type: string
      cfo_im_net_operating_cash_flow:
        type: string
      cfo_im_operating_cash_flow:
        type: string
      effect_of_exchange_rate_change:
        type: string
      file_id:
        type: string
      id:
        type: integer
      net_increase:
        type: string
      statement_id:
        type: string
    type: object
...
  models.FinancialStatement:
    properties:
      created_at:
        type: string
      currency:
        type: string
      employees:
        type: string
      file_id:
        type: string
      id:
        type: integer
      legal_entity_registration_number:
        type: string
      rounded_to_nearest:
        type: string
      source_schema:
        type: string
      source_type:
        type: string
      year:
        type: string
      year_ended_on:
        type: string
      year_started_on:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server for managing data based on the provided schema.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Your API Title
  version: "1.0"
paths:
...
  /cash-flow-statements:
    get:
      description: Retrieve a list of all cash flow statement records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CashFlowStatement'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get all cash flow statement entries
      tags:
      - cash-flow-statements
    post:
      consumes:
      - application/json
      description: Add a new cash flow statement record to the database
      parameters:
      - description: Cash Flow Statement data to create
        in: body
        name: cashFlowStatement
        required: true
        schema:
          $ref: '#/definitions/models.CashFlowStatement'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.CashFlowStatement'
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Create a new cash flow statement entry
      tags:
      - cash-flow-statements
  /cash-flow-statements/{id}:
    delete:
      description: Remove a cash flow statement record from the database using its
        ID
      parameters:
      - description: Cash Flow Statement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Cash Flow Statement not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Delete a cash flow statement entry by ID
      tags:
      - cash-flow-statements
    get:
      description: Retrieve details of a specific cash flow statement record using
        its ID
      parameters:
      - description: Cash Flow Statement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CashFlowStatement'
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Cash Flow Statement not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get a single cash flow statement entry by ID
      tags:
      - cash-flow-statements
    put:
      consumes:
      - application/json
      description: Modify the details of an existing cash flow statement record by
        ID
      parameters:
      - description: Cash Flow Statement ID
        in: path
        name: id
        required: true
        type: integer
      - description: Cash Flow Statement data to update
        in: body
        name: cashFlowStatement
        required: true
        schema:
          $ref: '#/definitions/models.CashFlowStatement'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CashFlowStatement'
        "400":
          description: Bad Request - Invalid ID format or input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Cash Flow Statement not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Update an existing cash flow statement entry
      tags:
      - cash-flow-statements
  /financial-statements:
    get:
      description: Retrieve a list of all financial statement records
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.FinancialStatement'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get all financial statement entries
      tags:
      - financial-statements
    post:
      consumes:
      - application/json
      description: Add a new financial statement record to the database
      parameters:
      - description: Financial Statement data to create
        in: body
        name: financialStatement
        required: true
        schema:
          $ref: '#/definitions/models.FinancialStatement'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.FinancialStatement'
        "400":
          description: Bad Request - Invalid input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Create a new financial statement entry
      tags:
      - financial-statements
  /financial-statements/{id}:
    delete:
      description: Remove a financial statement record from the database using its
        ID
      parameters:
      - description: Financial Statement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Financial Statement not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Delete a financial statement entry by ID
      tags:
      - financial-statements
    get:
      description: Retrieve details of a specific financial statement record using
        its ID
      parameters:
      - description: Financial Statement ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FinancialStatement'
        "400":
          description: Bad Request - Invalid ID format
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Financial Statement not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Get a single financial statement entry by ID
      tags:
      - financial-statements
    put:
      consumes:
      - application/json
      description: Modify the details of an existing financial statement record by
        ID
      parameters:
      - description: Financial Statement ID
        in: path
        name: id
        required: true
        type: integer
      - description: Financial Statement data to update
        in: body
        name: financialStatement
        required: true
        schema:
          $ref: '#/definitions/models.FinancialStatement'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.FinancialStatement'
        "400":
          description: Bad Request - Invalid ID format or input data
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "404":
          description: Not Found - Financial Statement not found
          schema:
            $ref: '#/definitions/handlers.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.HTTPError'
      summary: Update an existing financial statement entry
      tags:
      - financial-statements
...
schemes:
- http
- https
swagger: "2.0"